Switch to abstract class for config
